+++ src/views/OtherView.vue
@@ -1,0 +1,7 @@
+<template>
+  <div>
+    <h1>OTHER</h1>
+  </div>
+</template>
+
+<script setup lang="ts"></script>
